+++ src/middleware.ts
@@ -9,12 +9,12 @@ export * from './middleware/csrf'
 export * from './middleware/etag'
 export * from './jsx'
 export * from './middleware/jsx-renderer'
-export { jwt } from './middleware/jwt'
+export { jwt } from './middleware/jwt/jwt'
 export * from './middleware/logger'
 export * from './middleware/method-override'
 export * from './middleware/powered-by'
-export * from './middleware/timing'
+export * from './middleware/timing/timing'
 export * from './middleware/pretty-json'
-export * from './middleware/secure-headers'
+export * from './middleware/secure-headers/secure-headers'
 export * from './middleware/trailing-slash'
 export * from './adapter/deno/serve-static'
